# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OriCMS API
  version: 2.0.0
  description: |
    Current OpenAPI snapshot for the main live OriCMS route families.

    This spec is intentionally curated rather than exhaustive. It focuses on the
    active project-based API model and the current agent gateway, and it avoids
    the older site-based and tier-era contracts that used to live in this file.

    Canonical routing model:
    - management API: `/api/v1/*`
    - project-scoped routes: `/api/v1/projects/:projectId/*`
    - delivery routes: `/api/v1/delivery/projects/:projectId/*`
    - agent gateway: `/api/v1/agent/v1/*`
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/projects/{projectId}/members:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Members]
      summary: List project members
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Members returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    post:
      tags: [Members]
      summary: Invite or add a human member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  $ref: '#/components/schemas/ProjectRole'
      responses:
        '201':
          description: Member or invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
  /api/v1/projects/{projectId}/members/agent:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Members]
      summary: Create an AI agent member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                role:
                  type: string
                  enum: [admin, editor, viewer]
                expiresInDays:
                  type: integer
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
  /api/v1/projects/{projectId}/content-types:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Content Types]
      summary: List content types
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Content types returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    post:
      tags: [Content Types]
      summary: Create a content type
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentTypeInput'
      responses:
        '201':
          description: Content type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
  /api/v1/projects/{projectId}/content-types/{typeId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TypeId'
    get:
      tags: [Content Types]
      summary: Get a content type
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Content type returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    put:
      tags: [Content Types]
      summary: Update a content type
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentTypeInput'
      responses:
        '200':
          description: Content type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
    delete:
      tags: [Content Types]
      summary: Delete a content type
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Content type deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AgentToken
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TypeId:
      name: typeId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true
    GenericSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
...
    ProjectRole:
      type: string
      enum: [owner, admin, editor, viewer]
    ContentTypeInput:
      type: object
      required: [name, plural, label, labelPlural, fields]
      properties:
        name:
          type: string
        plural:
          type: string
        label:
          type: string
        labelPlural:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
...